address comment and change token arguments from u64 to u128

+++ programs/drift/src/state/lp_pool.rs
@@ -1,6 +1,6 @@
 use crate::error::{DriftResult, ErrorCode};
 use crate::math::casting::Cast;
-use crate::math::constants::{PERCENTAGE_PRECISION_I64, PRICE_PRECISION, PRICE_PRECISION_I64};
+use crate::math::constants::{PERCENTAGE_PRECISION_I64, PRICE_PRECISION_I64};
 use crate::math::safe_math::SafeMath;
 use crate::math::spot_balance::get_token_amount;
 use anchor_lang::prelude::*;
@@ -134,8 +134,8 @@ impl LPPool {
         out_constituent: &Constituent,
         in_spot_market: &SpotMarket,
         out_spot_market: &SpotMarket,
-        in_token_balance: u64,
-        out_token_balance: u64,
+        in_token_balance: u128,
+        out_token_balance: u128,
         in_target_weight: i64,
         out_target_weight: i64,
         in_amount: u64,
@@ -178,7 +178,7 @@ impl LPPool {
         oracle_map: &mut OracleMap, // might not need oracle_map depending on how accounts are passed in
         constituent: &Constituent,
         spot_market: &SpotMarket,
-        token_balance: u64,
+        token_balance: u128,
         amount: u64,
         target_weight: i64,
     ) -> DriftResult<i64> {
@@ -285,7 +285,7 @@ impl Size for Constituent {
 impl Constituent {
     /// Returns the full balance of the Constituent, the total of the amount in Constituent's token
     /// account and in Drift Borrow-Lend.
-    pub fn get_full_balance(&self, token_balance: u64) -> DriftResult<i128> {
+    pub fn get_full_balance(&self, token_balance: u128) -> DriftResult<i128> {
         match self.spot_balance.balance_type() {
             SpotBalanceType::Deposit => token_balance
                 .cast::<i128>()?
@@ -301,7 +301,7 @@ impl Constituent {
     pub fn get_weight(
         &self,
         price: i64,
-        token_balance: u64,
+        token_balance: u128,
         token_amount_delta: i64,
         lp_pool_aum: u128,
     ) -> DriftResult<i64> {
